+++ example/example.dart
@@ -1,4 +1,5 @@
 // import the core of the framework:
+
 import 'package:objd/core.dart';
 // import the custom pack:
 import './packs/examplePack.dart';