+++ lib/src/annotations/pack.dart
@@ -1,4 +1,5 @@
 /// The `@Pck()` annotation works similar to @Func. You annotate a File List variable and it generates a Widget for this Pack.
+
 class Pck {
   final String? name;
   final String? main;